Able to check profit

+++ src/main/java/Accessories/Accessory.java
@@ -4,7 +4,7 @@ import shop.Stock;
 
 public class Accessory extends Stock {
 
-    public Accessory(String brand, int boughtPrice, int sellPrice) {
+    public Accessory(String brand, Double boughtPrice, Double sellPrice) {
         super(brand, boughtPrice, sellPrice);
     }
 }
