+++ src/main/java/Instruments/Guitar.java
@@ -7,7 +7,7 @@ public class Guitar extends Instrument implements IPlay {
     private String shape;
 
 
-    public Guitar(String brand, String material, String colour, InstrumentType type, String shape, int boughtPrice, int sellPrice) {
+    public Guitar(String brand, String material, String colour, InstrumentType type, String shape, Double boughtPrice, Double sellPrice) {
         super(brand, material, colour, type, boughtPrice, sellPrice);
         this.shape = shape;
     }
