+++ src/main/java/shop/Stock.java
@@ -3,20 +3,20 @@ package shop;
 public abstract class Stock {
 
     private String brand;
-    private int boughtPrice;
-    private int sellPrice;
+    private Double boughtPrice;
+    private Double sellPrice;
 
-    public Stock(String brand, int boughtPrice, int sellPrice) {
+    public Stock(String brand, Double boughtPrice, Double sellPrice) {
         this.brand = brand;
         this.boughtPrice = boughtPrice;
         this.sellPrice = sellPrice;
     }
 
-    public int getBoughtPrice() {
+    public Double getBoughtPrice() {
         return boughtPrice;
     }
 
-    public int getSellPrice() {
+    public Double getSellPrice() {
         return sellPrice;
     }
